% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_rate.R
\name{activity_rate_distcorr}
\alias{activity_rate_distcorr}
\title{Distance corrected activity rate}
\usage{
activity_rate_distcorr(
  ds_model = NULL,
  obs_size,
  survey_mins,
  mean_event_size = mean(obs_size, na.rm = TRUE),
  sd_event_size = sd(obs_size, na.rm = TRUE),
  wilson_correction = TRUE,
  eda = NULL
)
}
\arguments{
\item{ds_model}{an object of class \code{dsmodel} fit using
\code{Distance::ds}. Must use metres as measure of distance}

\item{obs_size}{integer; vector of the number of individuals seen in each
observation. For no observations send zero length vector.}

\item{survey_mins}{numeric; vector holding the duration of each survey.
NAs ignored. Must use units of minutes.}

\item{mean_event_size}{numeric; Allows the user to submit a custom variable
for the mean event size.  Defaults to the mean of obs_size.}

\item{sd_event_size}{numeric; Allows the user to submit a custom variable
for the sd event size.  Defaults to the sd of obs_size.}

\item{wilson_correction}{boolean;  Apply wilson correction if there are
no observations? Defaults to true}

\item{eda}{optional numeric; Allows you to manually specify the EDA, must be in square metres. Defaults to \code{NULL} (extracting it from the distance model).}
}
\value{
list; the mean and sd of activity rate
}
\description{
The activity rate is defined as the number of flights per
square kilometer per year, assuming constant presence (we deal with that
at a later step).
}
\details{
Given a set of observations, survey durations and a related distance model
(fit using the \code{Distance} package), returns the activity rate in
individuals / year / km^2.  The optional Wilson Correction can be used to
generate an activity rate in the case when no observations are made of the
species during the survey run.
}
\examples{

obs <- c(0, 1, 1, 3, 4)
survey <- c(60, 60, 60, NA, 30)

activity_rate_distcorr(ds_example,
    obs_size = obs,
    survey_mins = survey
)

# zero observations using wilson correction

obs <- integer()
survey <- c(60, 60, 60, 45, 30)
activity_rate_distcorr(ds_example,
    obs_size = obs,
    survey_mins = survey,
    wilson_correction = TRUE
)

# zero observations without wilson correction returns NaN
# and warning

activity_rate_distcorr(ds_example,
    obs_size = obs,
    survey_mins = survey,
    wilson_correction = FALSE
)

}
\references{
Wilson score interval example:
\url{https://en.wikipedia.org/wiki/Binomial_proportion_confidence_interval#Wilson_score_interval}
}
