% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_rate_raptor}
\alias{activity_rate_raptor}
\title{Example activity rate}
\format{
An object of class \code{list} of length 2.
}
\usage{
activity_rate_raptor
}
\description{
Example output from \code{collision::activity_rate_distcorr} using the example survey, observation and raptor data. Used for examples vignetts
}
\details{
#' @docType data
}
\keyword{datasets}
