% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_num_bounds}
\alias{check_num_bounds}
\title{check input within bounds}
\usage{
check_num_bounds(x, min = Inf, max = Inf)
}
\arguments{
\item{x}{the input to check}

\item{min}{optional lower bound}

\item{max}{optional upper bound}
}
\description{
check input within bounds
}
