% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correction.R
\name{cluster_correction_a}
\alias{cluster_correction_a}
\title{Correct for turbine clusters}
\usage{
cluster_correction_a(df_turbines, eff_detection_width)
}
\arguments{
\item{df_turbines}{data.frame; a data.frame with one row per turbine,
containing the coordinates of each turbine in WGS 84}

\item{eff_detection_width}{numeric; the effective detection width,
which is usually 2 x effective detection radius.}
}
\value{
numeric; the cluster correction factor - the average number
of turbines per EDA (effective detection area).
}
\description{
TODO description
}
\details{
If the turbine layout is more compact and/or your EDR is large, then the
effective detection area contains more than one turbine.
The observed flux needs to be weighted to account for the effective turbines
in the observed area. This accounts for the fact that an observed flight can
only interact with one turbine at any given time. TODO
}
