% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correction.R
\name{cluster_correction_l}
\alias{cluster_correction_l}
\title{Correct for turbine distance (offshore)}
\usage{
cluster_correction_l(avg_min_distance, eff_detection_width)
}
\arguments{
\item{avg_min_distance}{numeric; the average distance between each turbine and its nearest neighbour}

\item{eff_detection_width}{numeric; the effective detection width,
which is usually 2 x effective detection radius.}
}
\value{
numeric; the (linear) cluster correction factor - the average number
of turbines per ESW (effective strip width).
}
\description{
TODO description
}
\details{
TODO
}
