\name{collision-package}
\alias{collision-package}
\alias{collision}
\docType{package}
\title{\packageTitle{collision}}
\description{\packageDescription{collision}}
\section{Package Content}{\packageIndices{collision}}
\author{\packageAuthor{collision}}
\section{Maintainer}{\packageMaintainer{collision}}
\keyword{package}
