% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helper.R
\name{data_from_distmodel}
\alias{data_from_distmodel}
\alias{w_from_distmodel}
\alias{edr_from_distmodel}
\alias{eda_from_distmodel}
\title{extract data from distance model}
\usage{
w_from_distmodel(ds_model)

edr_from_distmodel(ds_model)

eda_from_distmodel(ds_model)
}
\arguments{
\item{ds_model}{an object of class \code{dsmodel} fit using \code{Distance::ds}}
}
\description{
Helper functions to extract data from a distance model object
}
\examples{

summary(ds_example) # example data
w_from_distmodel(ds_example) # extract truncation parameter, w
edr_from_distmodel(ds_example) # extract effective detection radius
eda_from_distmodel(ds_example) # extract effective detection area

}
\references{
Miller DL, Rexstad E, Thomas L, Marshall L, Laake JL (2019).
"Distance Sampling in R." Journal of Statistical Software,
89(1), 1–28. doi: 10.18637/jss.v089.i01.
}
