% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_inputs.R
\name{define_bird}
\alias{define_bird}
\title{make a list of bird parameters for input into CRM}
\usage{
define_bird(species, bird_length, bird_speed)
}
\arguments{
\item{species}{char; species identifier}

\item{bird_length}{Length of archetype fixedbird (metres), or distribution
information using \link{set_random}}

\item{bird_speed}{numeric; Average flight speed (m/sec), or
distribution information using \link{set_random}}

\item{prop_day}{numeric; Proportion of a 24 hour day that species is active. 
A single number or distribution information using \link{set_random}.}

\item{obs_scale_factor}{numeric; A scale factor (0,1] for the proportion of
the flight activity contained within the daily
observation times. A single number or distribution 
information using \link{set_random}. Default 1.}
}
\value{
list where items define the value or stochastic distribution f
   or each parameter
}
\description{
Makes a list of avian biological parameters for the bird(s) being modelled
as a function directly in R. This function is designed for R-based workflows.
A function define_bird_csv can be used to bulk upload parameters for
multiple birds.
}
\details{
# Details

Each input to this function must be either

 - an object of the \code{class} \code{randInput}.
   See \link{set_random} for more information on setting
   up stochastic inputs.
 - a single number representing the average (preferably median) of the
   variable
}
\examples{
lst_bird <- define_bird(
  species = 'biggus raptorus',
  bird_length = set_random("rnorm", mean = 90, sd = 5), # (metres)
  bird_speed = set_random("rnorm", mean = 16.67, sd = 3) # m/s
)

}
