% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_obs}
\alias{df_obs}
\title{Example observations}
\format{
An object of class \code{data.frame} with 120 rows and 6 columns.
}
\usage{
df_obs
}
\description{
Data set of observations (with corresponding survey_ids) that mimic point count
observations of raptors. Used for examples and to construct \code{ds_raptor}
}
\details{
#' @docType data
}
\keyword{datasets}
