% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_obs_survey}
\alias{df_obs_survey}
\title{Example survey and observation}
\format{
An object of class \code{data.frame} with 120 rows and 9 columns.
}
\usage{
df_obs_survey
}
\description{
Combined data set of survey and observations that mimic point count 
observations of raptors. Used for examples and to construct \code{ds_raptor}
}
\details{
#' @docType data
}
\keyword{datasets}
