% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_survey}
\alias{df_survey}
\title{Example survey}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
df_survey
}
\description{
Data set of survey metadata that mimic point count
surveys for raptors. Used for examples.
}
\details{
#' @docType data
}
\keyword{datasets}
