% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_example}
\alias{ds_example}
\title{An object of class \code{dsmodel} fit using \code{Distance::ds}
The example model is fit to the data \code{book.tee.data} from \code{mrds}}
\format{
An object of class \code{dsmodel} of length 3.
}
\usage{
ds_example
}
\description{
#' @docType data
}
\keyword{datasets}
