% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_raptor}
\alias{ds_raptor}
\title{Distance model example - raptor}
\format{
An object of class \code{dsmodel} of length 3.
}
\usage{
ds_raptor
}
\description{
An object of class \code{dsmodel} fit using \code{Distance::ds}
The example model is fit to the data \code{df_obs_survey} from this package
}
\details{
#' @docType data
}
\keyword{datasets}
