% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux_per_year}
\alias{flux_per_year}
\title{Flux though the turbine area per year}
\usage{
flux_per_year(flux_per_min, flight_hours, months_on_site)
}
\arguments{
\item{flux_per_min}{flux through turbine in flights / min}

\item{flight_hours}{hours per day the species is active onsite}

\item{months_on_site}{months per year the species is present onsite}
}
\description{
Converts flights per minute from [flight_flux()] into an annual rate
}
\details{
Note this calculation assumes the \code{flux_per_min} is representative of the annual average flight rate. If the observation period does not represent the flights during active hours of the day and annual activity, the inputs below may need adjusting, or the annual flux may need to be calculated using an alternative method.
}
