% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.R
\name{lonlat_to_utm}
\alias{lonlat_to_utm}
\title{lon lat to UTM}
\usage{
lonlat_to_utm(lonlat)
}
\arguments{
\item{lonlat}{a lon, lat coordinate pair}
}
\description{
Equation from https://r.geocompx.org/reproj-geo-data#which-crs
}
