% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_flux.R
\name{obs_flux}
\alias{obs_flux}
\title{Calculate observed flights / time / area from point transects}
\usage{
obs_flux(encounter_rate, eff_detection_width, eff_detection_height)
}
\arguments{
\item{encounter_rate}{numeric; number of flights observed per unit time of survey
as output by \code{\link[=encounter_rate]{encounter_rate()}} or similar.}

\item{eff_detection_width}{numeric; Allows you to manually specify the effective detection width,
which is usually 2 x effective detection radius or 2 x effective strip (half) width.
Must be in the same units as \code{eff_detection_height}.}

\item{eff_detection_height}{numeric; The effective detection height. Because the
density of flights typically decreases with increasing height, the assumption of uniform
density required for traditional distance correction doesn't hold. So the observer's
detection function is effectively convolved with the flight height distribution
of the bird. Therefore, in most cases we recommend using the maximum tip height of the turbine
as the "effective detection height" and desktop truncating the observations to
that height. This is the simplest way to avoid artificially inflating or deflating
the flux through the turbine. It is possible to use another method to estimate
an effective detection height but we leave this to the analyst's judgement.
Must be in the same units as \code{eff_detection_width}.}
}
\value{
numeric; number of flights through a vertical plane of unit area per unit time.
Area units are defined by the units of the width and height, and time interval
is the same as referenced by the \code{encounter_rate} input.
}
\description{
This function converts the uncorrected encounter rate, the effective detection radius
(see for example \insertCite{Buckland2001;textual}{collision}) and the effective detection height into an estimate of flight flux
per unit time per unit area of vertical airspace \insertCite{Smales2013}{collision}.
}
\details{
\subsection{Observed flux from point counts}{

Given an encounter rate (\code{\link[=encounter_rate]{encounter_rate()}}), a distribution of flight heights
and a related distance model (fit using the \code{Distance} package \insertCite{Miller019a}{collision})
we calculate the flight flux through a rectangle with a width of 2 x (effective detection radius)
and a height equal to the effective detection height (typically set to the tip height of the proposed turbine)
in one unit of observation time.

Note the function accepts only single valued inputs, so stochastic inputs must
be sampled from prior to calling this function. For more information,
see \code{vignette("simple-simulation-example", package = "collision")}.
}
}
\examples{

## A simple example of calculating flux from a point count and
## using this to generate the number of flights through a turbine 
## in a year
## 
## # Step by step
## 

# four surveys
# Note the observation data should include observations within max rotor swept height
df_obs <- data.frame(
  size = c(0, 2 , 3, 0), #individuals observed per survey (below max rotor swept height)
  survey_duration = c(20, 20, 18, 20), # minutes
  # Optional survey weights to deal with stratification etc
  survey_weight = c(1,1,1,1)
)

rotor_diameter <- 300
hub_height <- 200
edr <- 800 # derive from distance model
max_rsh <- hub_height + rotor_diameter/2

# flights observed per minute of survey
flights_per_min <- encounter_rate(
  df_obs_summary = df_obs,
  wilson_correction = TRUE # Default
)

# observed flights through vertical plane of one metre squared in one minute
flights_per_m2_per_min <- obs_flux(
  encounter_rate = flights_per_min,
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh
)

# scale to turbine width and height
flights_turbine_min <- turbine_flights(
  obs_flux = flights_per_m2_per_min,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height
)

# scale to annual flights through area of rotor_diameter x turbine height 
flights_turbine_year <- flights_per_year(
  flights_per_time = flights_turbine_min,
  time_units = "min", # Default
  prop_day = 0.5,  #diurnal species
  prop_year = 1    # present all year
)

## Alternate calc using a wrapper function
## to go from observations to turbine flights per year
## 
flights_turbine_year2 <- turbine_flights_year(
  survey_type = "point", # only supported option currently
  encounter_rate = flights_per_min,
  time_units = "min",
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_day = 0.5,  #diurnal species
  prop_year = 1   # present all year
)

#They are the same
flights_turbine_year
flights_turbine_year2
}
\references{
\insertAllCited{}

\insertAllCited{}
}
\seealso{
\code{\link[=encounter_rate]{encounter_rate()}}

\code{\link[=turbine_flights]{turbine_flights()}} and \code{\link[=flights_per_year]{flights_per_year()}} for methods to expand the
observer flux into the flights through the turbine plane
}
