% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_coll_band_basic.R
\name{p_collision_band}
\alias{p_collision_band}
\title{Probability of collision given interaction from Band}
\usage{
p_collision_band(
  max_chord,
  pitch_deg,
  rotor_diam,
  rpm,
  bird_length,
  bird_wing_span,
  bird_flapping,
  bird_speed
)
}
\arguments{
\item{max_chord}{numeric; Maximum width of the turbine blade (metres).
Must be a static value}

\item{pitch_deg}{numeric; Pitch of the turbine blades (degrees).
Must be a static value}

\item{rotor_diam}{numeric; The length from turbine blade tip to blade 
tip (metres). Must be a static value}

\item{rpm}{Rotation speed (RPM).}

\item{bird_length}{numeric; Length of archetype bird (metres). Must be a
static value}

\item{bird_wing_span}{numeric; Wingspan of archetype bird (metres).
Must be a static value}

\item{bird_flapping}{numeric; 0 = bird_flapping, 1 = Soaring.
Must be a static value}

\item{bird_speed}{numeric; Average flight speed (m/sec).
Must be a static value}
}
\value{
numeric; Number between 0 and 1 representing the probability
of collision given interaction
}
\description{
Uses Band 2012 option 1. Simple implementation for testing and comparison.
}
\details{
Calculate probability of collision given interaction


This function calculates the probability of collision of a single
flight with a single turbine, given that the flight has
interacted with the turbine (i.e. \eqn{P(C|I)}). \cr
Formerly called \code{fun_bandWTG}.

Relies on the size and speed parameters for the bird and chosen
turbine model provided as a list.The list can be
produced using \code{make_lst_band()}.

Uses Band Collision Risk Model (see Band 2007 / Band 2012) and is
a direct reproduction of the publicly available Excel spreadsheet models.
}
\examples{


p_collision <- p_collision_band(
  max_chord = 3.51,
  pitch_deg = 20, 
  rotor_diam = 91.6,
  rpm = 16.1,
  bird_length = 1.1,
  wing_span = 2.0,
  flapping = 0,
  bird_speed = 6
)

}
