% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{parameterise_lnorm}
\alias{parameterise_lnorm}
\title{Given a mean and standard deviation, solves for log normal distribution
parameters meanlog and sdlog.}
\usage{
parameterise_lnorm(mean, sd)
}
\arguments{
\item{mean}{the mean}

\item{sd}{the standard deviation}
}
\value{
\code{list} object with the mean and standard deviation of the
distribution on the log scale. These are the \code{rlnorm} parameters \code{meanlog}
and \code{sdlog} used in the \code{\link[stats:rlnorm]{stats::rlnorm()}} function
}
\description{
Given a mean and standard deviation, solves for log normal distribution
parameters meanlog and sdlog.
}
\examples{

log_rate <- parameterise_lnorm(mean = 0.002, sd = 0.0001)

flight_flux_min <- set_random("rlnorm", 
                            meanlog = log_rate$mean, 
                            sdlog = log_rate$sd)
sample_input(flight_flux_min)

}
\seealso{
\code{\link[stats:rlnorm]{stats::rlnorm()}}
}
