% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_collision.R
\name{prob_collision_static}
\alias{prob_collision_static}
\title{Probability of collision with static turbine}
\usage{
prob_collision_static(
  d_base,
  d_rotormin,
  d_top,
  hh,
  blade_length,
  max_nac_h,
  max_nac_l,
  max_width_nacelle,
  rotor_diam,
  tilt_deg,
  max_chord,
  min_chord,
  blade_thickness_wide,
  blade_thickness_narrow,
  prop_at_height,
  prop_below_height,
  edr
)
}
\arguments{
\item{d_base}{diameter of tower at base}

\item{d_rotormin}{diameter of tower at base of rotor}

\item{d_top}{diameter of tower at top}

\item{hh}{hub height}

\item{blade_length}{blade length}

\item{max_nac_h}{"Maximum L (m)"  max nacelle length (side view)}

\item{max_nac_l}{"Maximum H (m)"  max nacelle height (side view)
#' @param max_width_nacelle "Maximum W and Hub Dia (m)" diam of nosecone == nacelle room width (back view)}

\item{tilt_deg}{Blade tilt in degrees}

\item{max_chord}{the chord of the blade at its widest point}

\item{min_chord}{the chord of the blade at its tip (thinnest point)}

\item{blade_thickness_wide}{the thickness of the blade (side on) at its widest point}

\item{blade_thickness_narrow}{the thickness of the blade (side on) at its thinnest point}

\item{prop_at_height}{proportion of flights at height}

\item{prop_below_height}{proportion of flights below rotor height}

\item{edr}{effective detection radius (m)}
}
\description{
Probability of collision calculated as the ratio of the sum of static presented
areas to the total airspace of rotor diameter and rotor swept height.
}
\details{
Presented areas are the sum of static  components, where
 static components are the tower below rotor height, tower above rotor height, 
 rotor and nacelle/nosecone
}
