% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{prob_interaction_flat}
\alias{prob_interaction_flat}
\title{Probability of interaction - flat utilisation}
\usage{
prob_interaction_flat(num_turbines)
}
\arguments{
\item{num_turbines}{the total number of turbines on the site (or sub site)
you are modelling}
}
\value{
numeric
}
\description{
TODO expand If the flight rate is not spatially variable, we can approximate
a flight by a random walk and calculate the average number of turbines
encountered by any flight is the square root of the total number of
turbines (See Smales et al 2013). It follows that the probability of
interation with a given turbine location is 1 / sqrt(number of turbines).
}
\examples{

prob_interaction_flat(num_turbines = 100)

}
