% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presented_area.R
\name{rotate_areas}
\alias{rotate_areas}
\title{Rotate areas}
\usage{
rotate_areas(front, side)
}
\arguments{
\item{front}{area of front-view in metres squared}

\item{side}{area of side-view in metres squared}
}
\value{
numeric
}
\description{
Rotate shape through 180 degrees. Internal function
}
\examples{
\dontrun{
rotate_areas(1,2)
}

}
