% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conversions.R
\name{rpm_to_radpersec}
\alias{rpm_to_radpersec}
\title{rpm to rad_per_sec}
\usage{
rpm_to_radpersec(rpm)
}
\arguments{
\item{rpm}{rotor speed in RPM}
}
\description{
Convert rotational speed in rpm to radians per second.
ROT rpm x (1/60sec) x (2pi rad) = ROT' rad / sec
}
\examples{
rpm_to_radpersec(1) # 0.1047
 
}
