% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sample.R
\name{sample_input}
\alias{sample_input}
\alias{sample_input.randInput}
\alias{sample_input.birdInput}
\alias{sample_input.turbineInput}
\alias{sample_input.default}
\title{Sample value from chosen distribution}
\usage{
sample_input(input, ...)

\method{sample_input}{randInput}(input, ..., n = 1)

\method{sample_input}{birdInput}(input, ..., n = 1)

\method{sample_input}{turbineInput}(input, ..., n = 1)

\method{sample_input}{default}(input, ..., n = 1)
}
\arguments{
\item{input}{an object of class \code{randInput} or \code{numeric}. See \link{set_random} for details}

\item{...}{Additional parameters passed to methods.}

\item{n}{integer. The number of samples to draw. Defaults to 1.}
}
\description{
Returns sampled value for random or fixed inputs
}
\examples{

var1 <- set_random("runif", min = 0, max = 10) # sample 1 (default)
is.randInput(var1) # should be true
sample_input(var1) # random
sample_input(var1, n=10) # samole 10 values

sample_input(17)  # fixed input
sample_input(17, n=10)  # fixed input, replicated

}
