% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sample.R
\name{set_random}
\alias{set_random}
\alias{is.randInput}
\title{Define how to sample a random input}
\usage{
set_random(distr, ...)

is.randInput(x)
}
\arguments{
\item{distr}{string. The name of the R function to return random deviates (e.g. \code{runif}, \code{rnorm}). See vignettes for examples of fitting to an empirical distribution.}

\item{...}{Additional parameters passed to the function \code{distr}. Do *not* set \code{n} (number of samples) here. This is set in \link{sample_input} to separate the simulation settings and statistical distribution definition.}

\item{x}{object to check}
}
\description{
Stochastic input definitions using the analysts choice of base
R sample distributions. Method to extend the \link{sample_input} function
}
\examples{

# sample from standard R functions
set_random("runif", min = 0, max = 10)
set_random("runif", n = 1, min = 0, max = 10) # n is ignored

set_random("rbeta", shape1 = 1, shape2 = 3, ncp = 0)
set_random("rgamma", shape = 1, rate = 1, scale = 1)

}
\seealso{
\link{sample_input} to sample a value according to the definition.
}
