% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_scaling.R
\name{turbine_flights}
\alias{turbine_flights}
\title{Convert observed flight flux to number of flights through turbine plane (i.e. interactions).}
\usage{
turbine_flights(obs_flux, spatial_correction = 1, rotor_diameter, hub_height)
}
\arguments{
\item{obs_flux}{numeric; number of flights through one unit area of vertical space
per unit time as output by \code{\link[=obs_flux]{obs_flux()}} or similar.}

\item{spatial_correction}{numeric; factor to correct flux for spatial considerations.
Either the cluster correction factor as output by \code{\link[=cluster_correction_a]{cluster_correction_a()}}/
\code{\link[=cluster_correction_l]{cluster_correction_l()}} or similar (assuming flat activity across the site)
or a correction based on the spatial distribution of flights on the site.
Defaults to 1.
TODO make better words}

\item{rotor_diameter}{numeric; rotor diameter. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m\eqn{^2}, \code{rotor_diameter} must be in m).}

\item{hub_height}{numeric; hub height. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m\eqn{^2}, \code{hub_height} must be in m).}
}
\value{
numeric; number of flights through turbine plane per unit time.
The turbine plane is a rectangle defined by the rotor diameter and tip height of the turbine.
Time interval is the same as referenced by \code{obs_flux} input.
}
\description{
Convert observed flight flux to number of flights through turbine plane (i.e. interactions).
}
\details{
\subsection{Scale to turbine}{

\code{\link[=turbine_flights]{turbine_flights()}} converts the flight flux through one unit area of vertical space
into the number of flights through an arbitrary vertical plane of width equal
to rotor diameter and height equal to the tip height of the turbine.

The time interval is not adjusted (i.e. if you input flights / min, it will
output flights / min).
}
}
\examples{

## A simple example of calculating flux from a point count and
## using this to generate the number of flights through a turbine 
## in a year
## 
## # Step by step
## 

# four surveys
# Note the observation data should include observations within max rotor swept height
df_obs <- data.frame(
  size = c(0, 2 , 3, 0), #individuals observed per survey (below max rotor swept height)
  survey_duration = c(20, 20, 18, 20), # minutes
  # Optional survey weights to deal with stratification etc
  survey_weight = c(1,1,1,1)
)

rotor_diameter <- 300
hub_height <- 200
edr <- 800 # derive from distance model
max_rsh <- hub_height + rotor_diameter/2

# flights observed per minute of survey
flights_per_min <- encounter_rate(
  df_obs_summary = df_obs,
  wilson_correction = TRUE # Default
)

# observed flights through vertical plane of one metre squared in one minute
flights_per_m2_per_min <- obs_flux(
  encounter_rate = flights_per_min,
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh
)

# scale to turbine width and height
flights_turbine_min <- turbine_flights(
  obs_flux = flights_per_m2_per_min,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height
)

# scale to annual flights through area of rotor_diameter x turbine height 
flights_turbine_year <- flights_per_year(
  flights_per_time = flights_turbine_min,
  time_units = "min", # Default
  prop_day = 0.5,  #diurnal species
  prop_year = 1    # present all year
)

## Alternate calc using a wrapper function
## to go from observations to turbine flights per year
## 
flights_turbine_year2 <- turbine_flights_year(
  survey_type = "point", # only supported option currently
  encounter_rate = flights_per_min,
  time_units = "min",
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_day = 0.5,  #diurnal species
  prop_year = 1   # present all year
)

#They are the same
flights_turbine_year
flights_turbine_year2
}
\seealso{
\code{\link[=obs_flux]{obs_flux()}}
}
