% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_wrapper.R
\name{turbine_flights_year}
\alias{turbine_flights_year}
\title{Estimate flights through turbine plane from observations}
\usage{
turbine_flights_year(
  survey_type = c("point"),
  encounter_rate,
  time_units = "min",
  eff_detection_width,
  eff_detection_height,
  spatial_correction = 1,
  rotor_diameter,
  hub_height,
  prop_day,
  prop_year
)
}
\arguments{
\item{survey_type}{character; type of survey. Currently just point transects
but methods for line transects and digital areal surveys are in development.}

\item{encounter_rate}{numeric; number of flights observed per unit time of survey
as output by \code{\link[=encounter_rate]{encounter_rate()}} or similar.}

\item{time_units}{Time units of \code{encounter_rate}. Defaults to "min" (i.e. flights per minute).}

\item{eff_detection_width}{numeric; Allows you to manually specify the effective detection width,
which is usually 2 x effective detection radius or 2 x effective strip (half) width.
Must be in the same units as \code{eff_detection_height}.}

\item{eff_detection_height}{numeric; The effective detection height. Because the
density of flights typically decreases with increasing height, the assumption of uniform
density required for traditional distance correction doesn't hold. So the observer's
detection function is effectively convolved with the flight height distribution
of the bird. Therefore, in most cases we recommend using the maximum tip height of the turbine
as the "effective detection height" and desktop truncating the observations to
that height. This is the simplest way to avoid artificially inflating or deflating
the flux through the turbine. It is possible to use another method to estimate
an effective detection height but we leave this to the analyst's judgement.
Must be in the same units as \code{eff_detection_width}.}

\item{spatial_correction}{numeric; factor to correct flux for spatial considerations.
Either the cluster correction factor as output by \code{\link[=cluster_correction_a]{cluster_correction_a()}}/
\code{\link[=cluster_correction_l]{cluster_correction_l()}} or similar (assuming flat activity across the site)
or a correction based on the spatial distribution of flights on the site.
Defaults to 1 (no correction).}

\item{rotor_diameter}{numeric; rotor diameter. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m\eqn{^2}, \code{rotor_diameter} must be in m).}

\item{hub_height}{numeric; hub height. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m\eqn{^2}, \code{hub_height} must be in m).}

\item{prop_day}{numeric; number between 0 and 1 representing the proportion
of a 24 hour day the species is active onsite. Also refer to the details below.}

\item{prop_year}{numeric; number between 0 and 1 representing the proportion
of a 12 month year the species is active onsite. Also refer to the details below.}
}
\description{
This is a wrapper to calculate the observed flight flux and scale
up to a turbine and a year.
}
\details{
Calculating the number of flights through a turbine
plane includes three steps
\itemize{
\item Calculate the observed flights in a vertical area from the
count of observed flights and details of the surveys, e.g. \code{\link[=obs_flux]{obs_flux()}}
\item Scale the observed flight flux to an area equivalent to the
rotor diameter by the maximum height of the turbine, e.g. \code{\link[=turbine_flights]{turbine_flights()}}
\item Scale the flights to the time period under study. There's a
helper function to scale to one year, \code{\link[=flights_per_year]{flights_per_year()}}
}
\subsection{Observed flux from point counts}{

Given an encounter rate (\code{\link[=encounter_rate]{encounter_rate()}}), a distribution of flight heights
and a related distance model (fit using the \code{Distance} package \insertCite{Miller019a}{collision})
we calculate the flight flux through a rectangle with a width of 2\eqn{\times}(effective detection radius)
and a height equal to the effective detection height (typically set to the tip
height of the proposed turbine) in one unit of observation time.

Note the function accepts only single valued inputs, so stochastic inputs must
be sampled from prior to calling this function. For more information,
see \code{vignette("simple-simulation-example", package = "collision")}.
}

\subsection{Scale to turbine}{

\code{\link[=turbine_flights]{turbine_flights()}} converts the flight flux through one unit area of vertical space
into the number of flights through an arbitrary vertical plane of width equal
to rotor diameter and height equal to the tip height of the turbine.

The time interval is not adjusted (i.e. if you input flights / min, it will
output flights / min).
}

\subsection{Scale to year}{

After obtaining the observed flux through a vertical area in one unit time,
we need to scale up to the relevant risk timeframe, often a year.

This can be done manually by the analyst for any time period, but we have
included a helper function for the case of scaling to one year (\code{\link[=flights_per_year]{flights_per_year()}}).
Note this function assumes the flux is the average flight flux when active onsite.
If surveys are conducted year round and the flux represents the annual average
then \code{prop_year} should be 1. If surveys are conducted only while the bird is
on site and the flux represents the average over the period the birds are on site
then \code{prop_year} should be the proportion of the year that the bird is on site.
For example, if a bird is onsite for three months of the year and the flux
was measured in that season only, then \code{prop_year = 3/12 = 0.25}.

Similarly care must be taken if the daily observation window does not overlap
completely with the birds activity. If the flight flux calculation
includes times when the species is not active the \code{prop_day} should be
adjusted to account for this, or the flux calculated only using
surveys during the activity period.
}
}
\examples{

## A simple example of calculating flux from a point count and
## using this to generate the number of flights through a turbine 
## in a year
## 
## # Step by step
## 

# four surveys
# Note the observation data should include observations within max rotor swept height
df_obs <- data.frame(
  size = c(0, 2 , 3, 0), #individuals observed per survey (below max rotor swept height)
  survey_duration = c(20, 20, 18, 20), # minutes
  # Optional survey weights to deal with stratification etc
  survey_weight = c(1,1,1,1)
)

rotor_diameter <- 300
hub_height <- 200
edr <- 800 # derive from distance model
max_rsh <- hub_height + rotor_diameter/2

# flights observed per minute of survey
flights_per_min <- encounter_rate(
  df_obs_summary = df_obs,
  wilson_correction = TRUE # Default
)

# observed flights through vertical plane of one metre squared in one minute
flights_per_m2_per_min <- obs_flux(
  encounter_rate = flights_per_min,
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh
)

# scale to turbine width and height
flights_turbine_min <- turbine_flights(
  obs_flux = flights_per_m2_per_min,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height
)

# scale to annual flights through area of rotor_diameter x turbine height 
flights_turbine_year <- flights_per_year(
  flights_per_time = flights_turbine_min,
  time_units = "min", # Default
  prop_day = 0.5,  #diurnal species
  prop_year = 1    # present all year
)

## Alternate calc using a wrapper function
## to go from observations to turbine flights per year
## 
flights_turbine_year2 <- turbine_flights_year(
  survey_type = "point", # only supported option currently
  encounter_rate = flights_per_min,
  time_units = "min",
  eff_detection_width = 2*edr,
  eff_detection_height = max_rsh,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_day = 0.5,  #diurnal species
  prop_year = 1   # present all year
)

#They are the same
flights_turbine_year
flights_turbine_year2
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=flights_per_year]{flights_per_year()}}, \code{\link[=turbine_flights]{turbine_flights()}}, \code{\link[=obs_flux]{obs_flux()}}
}
