% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_collision.R
\name{vertical_flux_plane_area}
\alias{vertical_flux_plane_area}
\title{Vertical flux plane area}
\usage{
vertical_flux_plane_area(hh, rotor_diam)
}
\arguments{
\item{hh}{hub height (m)}

\item{rotor_diam}{rotor diameter (m)}
}
\value{
area of plane of flight in m^2
}
\description{
The probability of collision given interaction requires the calculation of a
vertical flux window through which flights occur.  The width is 2x the
effective detection radius and the height is the max. height of the turbine
}
\examples{

vertical_flux_plane_area(
  hh = 150,
  rotor_diam = 210
)

}
