% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wte}
\alias{wte}
\title{Wedge-tailed Eagle data}
\format{
A object of class 'list' and custom class 'birdInput'
}
\usage{
wte
}
\description{
A non-stochastic example of bird data for vignettes and testing
Based off the Australian / Tasmanian Wedge-tailed Eagle
}
\details{
Created with [define_bird()]
}
\keyword{datasets}
