% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_inputs.R
\name{define_bird}
\alias{define_bird}
\title{make a list of bird parameters for input into CRM}
\usage{
define_bird(
  species,
  bird_length,
  bird_speed,
  prop_day,
  prop_year,
  avoidance_dynamic,
  avoidance_static
)
}
\arguments{
\item{species}{char; species identifier}

\item{bird_length}{Length of archetype bird (metres), or distribution
information using \link{set_random}}

\item{bird_speed}{numeric; Average flight speed (m/sec), or
distribution information using \link{set_random}}

\item{prop_day}{proportion of the day the bird is active onsite}

\item{prop_year}{proportion of the year (relative to the time period of the survey)
the bird is active onsite}

\item{avoidance_dynamic}{avoidance rate relative to the leading edge of the blade}

\item{avoidance_static}{avoidance rate relative to a static turbine}
}
\value{
list where items define the value or stochastic distribution f
   or each parameter
}
\description{
Makes a list of avian biological parameters for the bird(s) being modelled
as a function directly in R. This function is designed for R-based workflows.
A function define_bird_csv can be used to bulk upload parameters for
multiple birds.
}
\details{
# Details

Each input to this function must be either

 - an object of the \code{class} \code{randInput}.
   See \link{set_random} for more information on setting
   up stochastic inputs.
 - a single number representing the average (preferably median) of the
   variable
}
\examples{
lst_bird <- define_bird(
  species = 'biggus raptorus',
  bird_length = set_random("rnorm", mean = 90, sd = 5), # (metres)
  bird_speed = set_random("rnorm", mean = 16.67, sd = 3), # m/s
  prop_day = 0.5,
  prop_year = set_random("runif", min = 0.5, max = 0.6),
  avoidance_dynamic = 0.95,
  avoidance_static = 0.999
)

}
