% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_rate.R
\name{encounter_rate}
\alias{encounter_rate}
\title{Calculate uncorrected observed flights / minute of survey from point transects}
\usage{
encounter_rate(df_obs_summary, survey_units = "min", wilson_correction = TRUE)
}
\arguments{
\item{df_obs_summary}{data.frame; a data.frame with one row per survey containing at least columns \code{size} and
\code{survey_duration}. It can also optionally include a column \code{survey_weight} if needed
to account for stratification etc. When NULL (the default) will
weight surveys equally.}

\item{survey_units}{character; units of \code{survey_duration}. Defaults to "min".
Will return a warning if not "min" or "minutes".}

\item{wilson_correction}{boolean;  Apply wilson correction if there are
no observations. Defaults to TRUE (TODO add citation or explanation of wilson correction).}
}
\value{
numeric; number of flights observed in one minute of survey.
}
\description{
This function converts observed flights (movements), survey duration into the "encounter rate" per minute of survey. This value is uncorrected for the observer's effective detection area \insertCite{Buckland1993}{collision}.
}
\details{
The steps to calculate the number of flights through a turbine
plan includes three steps
\itemize{
\item Calculate the observed flights in a vertical square-metre area from the
count of observed flights and details of the surveys. See for example
\code{\link[=obs_flux]{obs_flux()}}
\item Scale the observed flight flux to an area equivalent to the
rotor diameter by the maximum height of the turbine, e.g. \code{\link[=turbine_flux]{turbine_flux()}}
\item Scale the observed flight flux to the time period under study. There's a
helper function to scale to one year e.g \code{\link[=flux_per_year]{flux_per_year()}}
}
\subsection{Observed flux from point counts}{

Given a set of observations, survey durations and a related distance model
(fit using the \code{Distance} package \insertCite{Miller019a}{collision}) we
calculate the flight flux through
a rectangle of width of 2*(effective detection radius) and height of the
observer's effective detection height (see TODO for more info about effective
detection height) in one minute of observation time.

The optional Wilson Correction can be used to
generate an flux in the case when no observations are made of the
species during the survey run.

Note the function accepts only single valued inputs, so stochastic inputs must
be sampled from prior to calling this function. See vignettes for more details
}

\subsection{Scale to turbine}{

\code{\link[=turbine_flux]{turbine_flux()}} converts the flight flux through one metre of vertical space
into the number of flights through an arbitrary vertical plane of width equal
to rotor diameter and height equal to the tip height of the turbine

The time interval is not adjusted (i.e. if you input flights / min, it will
output flights / min).
}

\subsection{Scale to year}{

After obtaining the observed flux through a vertical area in a minute,
we need to scale up to the relevant risk timeframe, often a year.

This can be done manually by the analyst for any time period, but we have
included a helper function for the case of scaling to one year (\code{\link[=flux_per_year]{flux_per_year()}}).
Note this function assumes the flux is the
average flight flux when active onsite.  If surveys are conducted year round
and the flux represents the annual average then \code{prop_year} should be 1. If
surveys are conducted only while the bird is on site and the flux represents
the average over the period the birds are on site then \code{prop_year} should be
the proportion of the year that the bird is on site.
For example, if a bird is onsite for three months of the year and the flux
was measured in season only, the \code{prop_year = 3/12 = 0.25}.

Similarly care must be taken if the daily observation window does not overlap
completely with the birds activity. If the flight flux calculation
includes times when the species is not active the \code{prop_day} should be
adjusted to account for this, or the flux calculated only using
surveys during the activity period.
}
}
\examples{

## A simple example of calculating flux from a point count and
## using this to generate the number of flights through a turbine 
## in a year
## 
## # Step by step
## 

df_obs <- data.frame(size = c(0, 2 , 3, 0), # four surveys
                     survey_duration = c(20, 20, 18, 20),
                     # Optional survey weights to deal with stratification etc
                     survey_weight = c(1,1,1,1))

prop_below_turbine_max <- 1
rotor_diameter <- 300
hub_height <- 200
edr <- 800
edh <- hub_height+0.5*rotor_diameter # can set to turbine height since all flights below turbine max

# flights observed per minute of survey
flights_per_min <- encounter_rate(
  df_obs_summary = df_obs,
  survey_units = "min", # Default
  wilson_correction = TRUE # Default
)

# flights through vertical plane of one metre squared in one minute
flights_per_m2_per_min <- obs_flux(
  encounter_rate = flights_per_min,
  eff_detection_width = edr,
  eff_detection_height = edh,
  width_units = "m", # Default
  height_units = "m" # Default
)

# scale to turbine width and height
flights_turbine_min <- turbine_flux(
  obs_flux = flights_per_m2_per_min,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_below_turbine_max = prop_below_turbine_max
)

# scale to annual flights through area of rotor_diameter x turbine height 
flights_turbine_year <- flux_per_year(
  flux_per_min = flights_turbine_min,
  prop_day = 0.5,  #diurnal species
  prop_year = 1    # present all year
)

## Alternate calc using a wrapper function
## to go from observations to turbine flights per year
## 
flights_turbine_year2 <- turbine_flux_year(
  survey_type = "point", # only supported option currently
  encounter_rate = flights_per_min,
  eff_detection_width = edr,
  eff_detection_height = edh,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_below_turbine_max = prop_below_turbine_max,
  prop_day = 0.5,  #diurnal species
  prop_year = 1,   # present all year
  width_units = "m", # Default
  height_units = "m" # Default
)

#They are the same
flights_turbine_year
flights_turbine_year2
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=turbine_flux]{turbine_flux()}} and \code{\link[=flux_per_year]{flux_per_year()}} for methods to expand the
observer flux into the flux through the turbine plane
}
