% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux_per_year}
\alias{flux_per_year}
\title{Flux though the turbine area per year}
\usage{
flux_per_year(flux_per_min, prop_day, prop_year)
}
\arguments{
\item{flux_per_min}{flux through turbine in flights / min}

\item{prop_day}{proportion of 24 hour day the species is active onsite}

\item{prop_year}{proportion of year the species is present onsite}
}
\value{
numeric; flux through turbine in flights / year
}
\description{
Converts flights per minute from [flight_flux()] into an annual rate
}
\details{
Note this calculation assumes the \code{flux_per_min} is representative of 
the annual average flight rate. If the observation period does not represent 
the flights during active hours of the day and annual activity, the inputs 
below may need adjusting, or the annual flux may need to be calculated using 
an alternative method.
}
\examples{

flux_per_year(
    flux_per_min = 100,
    prop_day = 0.5,
    prop_year = 1
)

}
