% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_collisions.R
\name{n_collision}
\alias{n_collision}
\title{Calculate annual collision rate}
\usage{
n_collision(avoidance_rate, n_interaction_yr, p_collision)
}
\arguments{
\item{avoidance_rate}{numeric; Number between 0 and 1 representing the
avoidance rate}

\item{n_interaction_yr}{numeric; the number of flights per year that pass
through the turbine disk. Calculated from [n_interaction()]}

\item{p_collision}{numeric; the probability of collision if an interaction occurs and avoidance is zero. Calculated from [p_collision()]}
}
\value{
numeric
}
\description{
Calculate annual collision rate
}
\examples{
n_collision(
  avoidance_rate = c(0.9, 0.99),
  n_interaction_yr = 100,
  p_collision = 0.01
)

}
