% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_collision.R
\name{prob_collision_dynamic}
\alias{prob_collision_dynamic}
\title{Probability of collision with leading edge of blade}
\usage{
prob_collision_dynamic(
  rpm,
  blade_length,
  max_width_nacelle,
  rotor_diam,
  blade_thickness_wide,
  blade_thickness_narrow,
  hh,
  bird_length,
  bird_speed,
  edr
)
}
\description{
TODO
}
