% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_collision.R
\name{prob_collision_dynamic}
\alias{prob_collision_dynamic}
\title{Probability of collision with leading edge of blade}
\usage{
prob_collision_dynamic(
  rpm,
  blade_length,
  max_width_nacelle,
  rotor_diam,
  blade_thickness_wide,
  blade_thickness_narrow,
  hh,
  bird_length,
  bird_speed,
  prop_at_height,
  prop_below_height
)
}
\arguments{
\item{rpm}{rotor speed (RPM)}

\item{blade_length}{blade length}

\item{max_width_nacelle}{"Maximum W and Hub Dia (m)" diam of nosecone == nacelle room width (back view)}

\item{rotor_diam}{rotor diameter (m)}

\item{blade_thickness_wide}{the thickness of the blade (side on) at its widest point}

\item{blade_thickness_narrow}{the thickness of the blade (side on) at its thinnest point}

\item{hh}{hub height}

\item{bird_length}{Length of archetype bird (metres)}

\item{bird_speed}{numeric; Average flight speed (m/sec)}

\item{prop_at_height}{proportion of flights at rotor swept height}

\item{prop_below_height}{proportion of flights below rotor swept height}
}
\value{
numeric; probability of collision with leading edge of blade.
Range from 0 to 1
}
\description{
Probability of collision with leading edge of blade
}
\examples{
# example code
prob_collision_dynamic(
 rpm = 15,
 blade_length = 60,
 max_width_nacelle = 4,
 rotor_diam = 124,
 blade_thickness_wide = 2.5,
 blade_thickness_narrow = 0.1,
 hh = 10,
 bird_length = 0.9,
 bird_speed = 12, 
 prop_at_height = 0.5,
 prop_below_height = 0.2
)


}
