% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_collision.R
\name{prob_collision_static}
\alias{prob_collision_static}
\title{Probability of collision with static turbine}
\usage{
prob_collision_static(
  d_base,
  d_rotormin,
  d_top,
  hh,
  blade_length,
  max_nac_h,
  max_nac_l,
  max_width_nacelle,
  rotor_diam,
  tilt_deg,
  max_chord,
  min_chord,
  blade_thickness_wide,
  blade_thickness_narrow,
  prop_at_height,
  prop_below_height
)
}
\arguments{
\item{d_base}{diameter of tower at base}

\item{d_rotormin}{diameter of tower at base of rotor}

\item{d_top}{diameter of tower at top}

\item{hh}{hub height}

\item{blade_length}{blade length}

\item{max_nac_h}{max nacelle length (side view)}

\item{max_nac_l}{max nacelle height (side view)}

\item{max_width_nacelle}{nacelle room width (back view)}

\item{rotor_diam}{rotor diameter (m)}

\item{tilt_deg}{Blade tilt in degrees}

\item{max_chord}{the chord of the blade at its widest point}

\item{min_chord}{the chord of the blade at its tip (thinnest point)}

\item{blade_thickness_wide}{the thickness of the blade (side on) at its widest point}

\item{blade_thickness_narrow}{the thickness of the blade (side on) at its thinnest point}

\item{prop_at_height}{proportion of flights at rotor swept height}

\item{prop_below_height}{proportion of flights below rotor swept height}
}
\value{
numeric; probability of collision with static turbine.
        Range from 0 to 1
}
\description{
Probability of collision calculated as the ratio of the sum of static presented
areas to the total airspace of rotor diameter and rotor swept height.
}
\details{
Presented areas are the sum of static  components, where
 static components are the tower below rotor height, tower above rotor height, 
 rotor and nacelle/nosecone
}
\examples{


# all in metres
prob_collision_static(
 d_base = 5.0,
 d_rotormin = 3.5,
 d_top = 3,
 hh = 10,
 blade_length = 60,
 max_nac_h = 4,
 max_nac_l = 13,
 max_width_nacelle = 4,
 rotor_diam = 124,
 tilt_deg = 6,
 max_chord = 3.5,
 min_chord = 0.4,
 blade_thickness_wide = 2.5,
 blade_thickness_narrow = 0.1,
 prop_at_height = 0.5,
 prop_below_height = 0.2
)



}
