% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_scaling.R
\name{turbine_flights}
\alias{turbine_flights}
\title{Convert observed flight flux to number flights through turbine plane
(the turbine plane is a rectangle defined by the rotor diameter and the maximum rotor swept height of the turbine).}
\usage{
turbine_flights(obs_flux, rotor_diameter, hub_height)
}
\arguments{
\item{obs_flux}{numeric; number of flights through one unit area of vertical space
per unit time as output by \code{\link[=obs_flux]{obs_flux()}} or similar}

\item{rotor_diameter}{numeric; rotor diameter. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m^2, \code{rotor_diameter} must be in m).}

\item{hub_height}{numeric; hub height. Must be in the equivalent units to
the unit area of \code{obs_flux} (i.e., if the \code{obs_flux} is per m^2, \code{hub_height} must be in m).}
}
\value{
numeric; number of flights through turbine plane per unit time.
Time interval is the same as referenced by \code{obs_flux} input.
}
\description{
Convert observed flight flux to number flights through turbine plane
(the turbine plane is a rectangle defined by the rotor diameter and the maximum rotor swept height of the turbine).
}
\details{
The steps to calculate the number of flights through a turbine
plane includes three steps
\itemize{
\item Calculate the observed flights in a vertical area from the
count of observed flights and details of the surveys. See for example
\code{\link[=obs_flux]{obs_flux()}}
\item Scale the observed flight flux to an area equivalent to the
rotor diameter by the maximum height of the turbine, e.g. \code{\link[=turbine_flights]{turbine_flights()}}
\item Scale the flights to the time period under study. There's a
helper function to scale to one year from minutes e.g \code{\link[=flights_per_year]{flights_per_year()}}
}
\subsection{Observed flux from point counts}{

Given a set of observations, survey durations, a distribution of flight heights
and a related distance model (fit using the \code{Distance} package \insertCite{Miller019a}{collision})
we calculate the flight flux through a rectangle of width of 2*(effective detection radius)
and mean (expected value) of the height distribution (because \eqn{E[X] = \int_{-\infty}^{\infty}xf(x)dx}
approximates the area in which we expect one flight if we make a simplifying assumption
of uniform vertical density) in one unit of observation time.

Note the function accepts only single valued inputs, so stochastic inputs must
be sampled from prior to calling this function. See vignettes for more details
}

\subsection{Scale to turbine}{

\code{\link[=turbine_flights]{turbine_flights()}} converts the flight flux through one unit area of vertical space
into the number of flights through an arbitrary vertical plane of width equal
to rotor diameter and height equal to the tip height of the turbine.

The time interval is not adjusted (i.e. if you input flights / min, it will
output flights / min).
}

\subsection{Scale to year}{

After obtaining the observed flux through a vertical area in a minute,
we need to scale up to the relevant risk timeframe, often a year.

This can be done manually by the analyst for any time period, but we have
included a helper function for the case of scaling to one year (\code{\link[=flights_per_year]{flights_per_year()}}).
Note this function assumes the flux is the
average flight flux when active onsite.  If surveys are conducted year round
and the flux represents the annual average then \code{prop_year} should be 1. If
surveys are conducted only while the bird is on site and the flux represents
the average over the period the birds are on site then \code{prop_year} should be
the proportion of the year that the bird is on site.
For example, if a bird is onsite for three months of the year and the flux
was measured in season only, the \code{prop_year = 3/12 = 0.25}.

Similarly care must be taken if the daily observation window does not overlap
completely with the birds activity. If the flight flux calculation
includes times when the species is not active the \code{prop_day} should be
adjusted to account for this, or the flux calculated only using
surveys during the activity period.
}
}
\examples{

## A simple example of calculating flux from a point count and
## using this to generate the number of flights through a turbine 
## in a year
## 
## # Step by step
## 

df_obs <- data.frame(size = c(0, 2 , 3, 0), # four surveys
                     survey_duration = c(20, 20, 18, 20),
                     # Optional survey weights to deal with stratification etc
                     survey_weight = c(1,1,1,1))

prop_below_turbine_max <- 1
rotor_diameter <- 300
hub_height <- 200
edr <- 800
edh <- hub_height+0.5*rotor_diameter # can set to turbine height since all flights below turbine max

# flights observed per minute of survey
flights_per_min <- encounter_rate(
  df_obs_summary = df_obs,
  survey_units = "min", # Default
  wilson_correction = TRUE # Default
)

# flights through vertical plane of one metre squared in one minute
flights_per_m2_per_min <- obs_flux(
  encounter_rate = flights_per_min,
  eff_detection_width = edr,
  eff_detection_height = edh,
  width_units = "m", # Default
  height_units = "m" # Default
)

# scale to turbine width and height
flights_turbine_min <- turbine_flux(
  obs_flux = flights_per_m2_per_min,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_below_turbine_max = prop_below_turbine_max
)

# scale to annual flights through area of rotor_diameter x turbine height 
flights_turbine_year <- flux_per_year(
  flux_per_min = flights_turbine_min,
  prop_day = 0.5,  #diurnal species
  prop_year = 1    # present all year
)

## Alternate calc using a wrapper function
## to go from observations to turbine flights per year
## 
flights_turbine_year2 <- turbine_flux_year(
  survey_type = "point", # only supported option currently
  encounter_rate = flights_per_min,
  eff_detection_width = edr,
  eff_detection_height = edh,
  rotor_diameter = rotor_diameter,
  hub_height = hub_height,
  prop_below_turbine_max = prop_below_turbine_max,
  prop_day = 0.5,  #diurnal species
  prop_year = 1,   # present all year
  width_units = "m", # Default
  height_units = "m" # Default
)

#They are the same
flights_turbine_year
flights_turbine_year2
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=obs_flux]{obs_flux()}}
}
