% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v90_single}
\alias{v90_single}
\title{Single turbine example data}
\format{
A object of class 'list' and custom class 'turbineInput':
}
\usage{
v90_single
}
\description{
An object of class \code{turbineInput} containing example turbine dimensions
based off a Vestas V90 example. This version has all fixed (i.e. no stochastic)
input values and is useful for vignettes and tests
}
\details{
Created with \code{\link[=define_turbine]{define_turbine()}}
}
\keyword{datasets}
